Remove outdated tests, move serialization_speed_test to a test case of Serialization_test.
